+++ Jellyfin.Xtream/Client/Models/Category.cs
@@ -21,11 +21,11 @@ namespace Jellyfin.Xtream.Client.Models;
 public class Category
 {
     [JsonProperty("category_id")]
-    public int CategoryId { get; set; }
-
-    [JsonProperty("category_name")]
-    public string CategoryName { get; set; } = string.Empty;
-
-    [JsonProperty("parent_id")]
-    public int ParentId { get; set; }
+    public long CategoryId { get; set; }
+ 
+     [JsonProperty("category_name")]
+     public string CategoryName { get; set; } = string.Empty;
+ 
+     [JsonProperty("parent_id")]
+     public long ParentId { get; set; }
 }
