+++ Jellyfin.Xtream/Client/Models/EpgInfo.cs
@@ -23,10 +23,10 @@ namespace Jellyfin.Xtream.Client.Models;
 public class EpgInfo
 {
     [JsonProperty("id")]
-    public int Id { get; set; }
-
-    [JsonProperty("epg_id")]
-    public int EpgId { get; set; }
+    public long Id { get; set; }
+ 
+     [JsonProperty("epg_id")]
+     public long EpgId { get; set; }
 
     [JsonConverter(typeof(Base64Converter))]
     [JsonProperty("title")]
